+++ CHANGELOG.md
@@ -16,6 +16,7 @@ For a full diff see [`2.0.0...3.0.0`][2.0.0...3.0.0].
 
 - Required [`ergebnis/json-pointer`](https://github.com/ergebnis/json-pointer) ([#195]), by [@localheinz]
 - Started throwing an `Exception\CanNotResolve` exception instead of an `Exception\ResolvedToRootSchema` when the `JsonPointer` is not a valid URI fragment identifier representation of a JSON pointer ([#202]), by [@localheinz]
+- Started using `Ergebnis\Json\Pointer\JsonPointer` instead of `Ergebnis\Json\SchemaValidator\JsonPointer` ([#200]), by [@localheinz]
 
 ### Removed
 
@@ -82,6 +83,7 @@ For a full diff see [`dcd4cfb...1.0.0`][dcd4cfb...1.0.0].
 [#169]: https://github.com/ergebnis/json-schema-validator/pull/169
 [#172]: https://github.com/ergebnis/json-schema-validator/pull/172
 [#195]: https://github.com/ergebnis/json-schema-validator/pull/195
+[#200]: https://github.com/ergebnis/json-schema-validator/pull/200
 [#202]: https://github.com/ergebnis/json-schema-validator/pull/202
 [#203]: https://github.com/ergebnis/json-schema-validator/pull/203